Rearchitecture project wroking

+++ domain/src/main/java/com/mupper/domain/LatLng.kt
@@ -5,4 +5,4 @@ package com.mupper.domain
  * Created by jesus.medina on 12/2019.
  * Mupper
  */
-data class LatLng(val latitude: Float, val longitude: Float)
+data class LatLng(val latitude: Double, val longitude: Double)
